feat: move to conf folder
