+++ .vscode/settings.json
@@ -1,4 +1,5 @@
 {
     "typescript.tsdk": "node_modules/typescript/lib",
-    "files.eol": "\n"
+    "files.eol": "\n",
+    "eslint.experimental.useFlatConfig": true // https://github.com/microsoft/vscode-eslint/issues/1644
 }
