+++ conf/browser.js
